requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
faiss-cpu==1.7.4
Flask==2.3.2
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-search-results==2.4.2
googleapis-common-protos==1.59.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
langchain==0.0.155
loguru==0.7.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
pexpect==4.8.0
Pillow==9.5.0
pinecone-client==2.2.1
pkgutil-resolve-name==1.3.10
platformdirs==3.5.0
protobuf==3.20.3
ptyprocess==0.7.0
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pypdf==3.8.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.29.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
streamlit==1.22.0
streamlit-chat==0.0.2.2
tenacity==8.2.2
tiktoken==0.4.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
tweepy==4.14.0
typing-extensions==4.5.0
typing-inspect==0.8.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
